+++ hrp/internal/env/env.go
@@ -22,15 +22,18 @@ var (
 const (
 	ResultsDirName     = "results"
 	ScreenshotsDirName = "screenshots"
+	ActionLogDireName  = "action_log"
 )
 
 var (
-	RootDir         string
-	ResultsDir      string
-	ResultsPath     string
-	ScreenShotsPath string
-	StartTime       = time.Now()
-	StartTimeStr    = StartTime.Format("20060102150405")
+	RootDir                 string
+	ResultsDir              string
+	ResultsPath             string
+	ScreenShotsPath         string
+	StartTime               = time.Now()
+	StartTimeStr            = StartTime.Format("20060102150405")
+	ActionLogFilePath       string
+	DeviceActionLogFilePath string
 )
 
 func init() {
@@ -43,4 +46,6 @@ func init() {
 	ResultsDir = filepath.Join(ResultsDirName, StartTimeStr)
 	ResultsPath = filepath.Join(RootDir, ResultsDir)
 	ScreenShotsPath = filepath.Join(ResultsPath, ScreenshotsDirName)
+	ActionLogFilePath = filepath.Join(ResultsDir, ActionLogDireName)
+	DeviceActionLogFilePath = "/sdcard/Android/data/io.appium.uiautomator2.server/files/hodor"
 }